refactored example
