+++ libsq/core/oncecache/example/hrsystem/README.md
@@ -1,0 +1,7 @@
+# hrsystem: oncecache example
+
+`hrsystem` contains a `oncecache` example. It is a simple staff directory backed by a
+pretend DB. It uses `oncecache` to reduce DB calls.
+
+One of the key things that `hrsystem` demonstrates is cache entry propagation across a
+composite cache.
