+++ libsq/core/oncecache/example/hrsystem/stats.go
@@ -1,4 +1,4 @@
-package staffdir
+package hrsystem
 
 import (
 	"fmt"
