--- libsq/core/oncecache/example/staffdir/README.md
@@ -1,4 +1,0 @@
-# staffdir: ocache example
-
-`staffdir` contains a `ocache` example. It is a simple staff directory backed by a DB.
-uses `oncecache` to cache the staff data.
